requirements.txt
# Core (install PyTorch separately for your CUDA version)
# See: https://pytorch.org/get-started/locally/
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.1
noisereduce
demucs==4.0.1
huggingface-hub>=0.28.1

# YingMusic-SVC dependencies
munch==4.0.0
einops==0.8.0
descript-audio-codec==1.0.0
resemblyzer
transformers==4.46.3
hydra-core==1.3.2
pyyaml
python-dotenv
ml_collections
loralib
beartype==0.14.1
rotary_embedding_torch==0.3.5
